Added password encryption on creating/editing admins

+++ public/manage_admins.php
@@ -5,7 +5,6 @@
 
 <?php $all_admins = find_all_admins();?>
 <div id="navigation">
-        <a href="create_admin.php"> Create Admin </a>
 </div>
 
 <div id="page">
@@ -19,12 +18,14 @@
         <?php while($admin = mysqli_fetch_assoc($all_admins)){
             $output  = "<tr>";
             $output .= "<td>{$admin['username']}</td>";
-            $output .= "<td><a href=\"edit_admin.php?id={$admin['id']}\">Edit</a>&nbsp;<a href=\"delete_admin.php?id={$admin['id']}\">Delete</a></td>";
+            $output .= "<td><a href=\"edit_admin.php?id={$admin['id']}\">Edit</a>&nbsp;<a href=\"delete_admin.php?id={$admin['id']}\" onclick=\"return confirm('Are you sure?')\">Delete</a></td>";
             $output .= "<td></td>";
             $output .= "<tr>";
             echo "$output";
         }?>
         </table>
+    <br/>
+    <a href="create_admin.php"> Create Admin </a>
 </div>
 
 <?php include("../includes/layouts/footer.html")?>
